Fixed #32915 -- Refactored runserver command to restore traceback
    The traceback for `runserver --nostatic` with broken/missing
    settings was hidden because the custom runserver command which
    added `--nostatic` to the command args was only instantiated if
    the settings had no errors.
