+++ tests/settings_tests/tests.py
@@ -338,6 +338,12 @@ class SettingsTests(SimpleTestCase):
         with self.assertRaisesMessage(ValueError, "Incorrect timezone setting: test"):
             settings._setup()
 
+    def test_import_non_existing_module(self):
+        exc = ModuleNotFoundError("No module named 'fake_module'", name="fake_module")
+        with mock.patch("importlib.import_module", side_effect=exc):
+            with self.assertRaisesMessage(ImportError, "No module named 'fake_module'"):
+                Settings("fake_settings_module")
+
 
 class TestComplexSettingOverride(SimpleTestCase):
     def setUp(self):
