+++ tests/staticfiles_tests/test_management.py
@@ -11,11 +11,11 @@ from admin_scripts.tests import AdminScriptTestCase
 
 from django.conf import STATICFILES_STORAGE_ALIAS, settings
 from django.contrib.staticfiles import storage
-from django.contrib.staticfiles.management.commands import collectstatic, runserver
+from django.contrib.staticfiles.management.commands import collectstatic
 from django.core.exceptions import ImproperlyConfigured
 from django.core.management import CommandError, call_command
 from django.core.management.base import SystemCheckError
-from django.test import RequestFactory, override_settings
+from django.test import override_settings
 from django.test.utils import extend_sys_path
 from django.utils._os import symlinks_supported
 from django.utils.functional import empty
@@ -31,27 +31,6 @@ class TestNoFilesCreated:
         Make sure no files were create in the destination directory.
         """
         self.assertEqual(os.listdir(settings.STATIC_ROOT), [])
-
-
-class TestRunserver(StaticFilesTestCase):
-    @override_settings(MIDDLEWARE=["django.middleware.common.CommonMiddleware"])
-    def test_middleware_loaded_only_once(self):
-        command = runserver.Command()
-        with mock.patch("django.middleware.common.CommonMiddleware") as mocked:
-            command.get_handler(use_static_handler=True, insecure_serving=True)
-            self.assertEqual(mocked.call_count, 1)
-
-    def test_404_response(self):
-        command = runserver.Command()
-        handler = command.get_handler(use_static_handler=True, insecure_serving=True)
-        missing_static_file = os.path.join(settings.STATIC_URL, "unknown.css")
-        req = RequestFactory().get(missing_static_file)
-        with override_settings(DEBUG=False):
-            response = handler.get_response(req)
-            self.assertEqual(response.status_code, 404)
-        with override_settings(DEBUG=True):
-            response = handler.get_response(req)
-            self.assertEqual(response.status_code, 404)
 
 
 class TestFindStatic(TestDefaults, CollectionTestCase):
